changed some fields in Stats.DTO and commented some methods
JavaDoc for methods and endpoints will be added as soon as I have a final version.
At the moment everything is stable and works just fine but more optimization can occur , on which I will focus next week , along with better code quality .
Special Tanks to Ana for helping me with SQL

+++ src/main/java/com/atoss/idea/management/system/repository/dto/StatisticsDTO.java
@@ -26,11 +26,11 @@ public class StatisticsDTO {
 
     private Long openIdeas;
 
-    private Long draftP;
+    private Double draftP;
 
-    private Long openP;
+    private Double openP;
 
-    private Long implP;
+    private Double implP;
 
     private List<IdeaResponseDTO> mostCommentedIdeas;
